requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
deep==0.10
defusedxml==0.7.1
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
iniconfig==1.1.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
ordered-set==4.0.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
peewee==3.14.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pkg_resources==0.0.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
PyLaTeX==1.4.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
pyzmq==22.2.1
requests==2.26.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
streamlit==0.88.0
terminado==0.11.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
tzlocal==3.0
urllib3==1.26.6
uuid==1.30
validators==0.18.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
